benchlab/vu/requirements.txt
requests>=2.31.0
PyYAML>=6.0
pyserial>=3.6
blessed>=1.20.0
Pillow>=10.0.0
windows-curses>=2.3.2; sys_platform == "win32"

